+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.16] - 2021-12-07
+### Added
+- Offline rendering for `plotly`
+
+### Changed
+- Requires `matplotlib>=3.5.0`
+
 ## [0.1.15] - 2021-12-07
 ### Changed
 - Requires `instancelib>=0.3.6.2`
@@ -130,6 +137,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved data wrappers from `text_explainability` to `genbase`
 
 [Unreleased]: https://git.science.uu.nl/m.j.robeer/genbase
+[0.1.16]: https://pypi.org/project/genbase/0.1.16/
+[0.1.15]: https://pypi.org/project/genbase/0.1.15/
 [0.1.14]: https://pypi.org/project/genbase/0.1.14/
 [0.1.13]: https://pypi.org/project/genbase/0.1.13/
 [0.1.12]: https://pypi.org/project/genbase/0.1.12/